+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.1.2] - 2015-02-17 12:30 GMT
+##### Fixed
+- Fix key $constraint_value on check method
+
+## [1.1.1] - 2015-02-17 12:30 GMT
+##### Fixed
+- Folder renamed for src provider.
+
+## [1.1.0] - 2015-02-17 12:30 GMT
+##### Changed
+- Support for get, set and save different rows of configuration with extra param constraint_value. - Cmarfil
 
 ## [1.0.2] - 2015-01-21 16:34 GMT
 ##### Changed
@@ -17,6 +28,7 @@ All notable changes to this project will be documented in this file.
 
 [#1]: https://github.com/Grimthorr/laravel-user-settings/issues/1
 
+[1.1.0]: https://github.com/cmarfil/laravel-multiuser-json-settings/compare/Grimthorr:master...master
 [1.0.2]: https://github.com/Grimthorr/laravel-user-settings/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/Grimthorr/laravel-user-settings/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/Grimthorr/laravel-user-settings/tree/1.0.0